Divided WorkshopCreatRequestDto class into 5 classes (WorkshopMainRequiredPropertiesDto, WorkshopRequiredPropertiesDto, WorkshopDescriptionDto, WorkshopContactsDto, WorkshopStaffDto) and added polymorphism between them.
