+++ OutOfSchool/OutOfSchool.BusinessLogic/Models/Workshops/Drafts/WorkshopContactsDto.cs
@@ -1,0 +1,26 @@
+﻿using System.ComponentModel.DataAnnotations;
+using Microsoft.AspNetCore.Mvc;
+using OutOfSchool.BusinessLogic.Util.JsonTools;
+
+namespace OutOfSchool.BusinessLogic.Models.Workshops.Drafts;
+public class WorkshopContactsDto : WorkshopDescriptionDto
+{
+    [DataType(DataType.Url)]
+    [MaxLength(Constants.MaxUnifiedUrlLength)]
+    public string Website { get; set; } = string.Empty;
+
+    [DataType(DataType.Url)]
+    [MaxLength(Constants.MaxUnifiedUrlLength)]
+    public string Facebook { get; set; } = string.Empty;
+
+    [DataType(DataType.Url)]
+    [MaxLength(Constants.MaxUnifiedUrlLength)]
+    public string Instagram { get; set; } = string.Empty;
+
+    [Required]
+    public long AddressId { get; set; }
+
+    [Required]
+    [ModelBinder(BinderType = typeof(JsonModelBinder))]
+    public AddressDto Address { get; set; }
+}
