+++ OutOfSchool/OutOfSchool.BusinessLogic/Models/Workshops/Drafts/WorkshopMainRequiredPropertiesDto.cs
@@ -1,0 +1,106 @@
+﻿using System.ComponentModel.DataAnnotations;
+using System.ComponentModel.DataAnnotations.Schema;
+using System.Text.Json.Serialization;
+using Microsoft.AspNetCore.Mvc;
+using OutOfSchool.BusinessLogic.Util.CustomValidation;
+using OutOfSchool.BusinessLogic.Util.JsonTools;
+using OutOfSchool.Common.Enums;
+using OutOfSchool.Common.Validators;
+using OutOfSchool.Services.Enums;
+
+namespace OutOfSchool.BusinessLogic.Models.Workshops.Drafts;
+
+[JsonDerivedType(typeof(WorkshopMainRequiredPropertiesDto), typeDiscriminator: "withMainProperties")]
+[JsonDerivedType(typeof(WorkshopRequiredPropertiesDto), typeDiscriminator: "withOtherRequiredProperties")]
+[JsonDerivedType(typeof(WorkshopDescriptionDto), typeDiscriminator: "withDescription")]
+[JsonDerivedType(typeof(WorkshopContactsDto), typeDiscriminator: "withContacts")]
+[JsonDerivedType(typeof(WorkshopStaffDto), typeDiscriminator: "withStaff")]
+public class WorkshopMainRequiredPropertiesDto : IValidatableObject //16 + 12 + 12 + 5 + 3 = 48
+{
+    public Guid Id { get; set; }
+
+    [Required(ErrorMessage = "Workshop title is required")]
+    [MinLength(Constants.MinWorkshopTitleLength)]
+    [MaxLength(Constants.MaxWorkshopTitleLength)]
+    public string Title { get; set; } = string.Empty;
+
+    [Required(ErrorMessage = "Workshop short title is required")]
+    [MinLength(Constants.MinWorkshopShortTitleLength)]
+    [MaxLength(Constants.MaxWorkshopShortTitleLength)]
+    public string ShortTitle { get; set; } = string.Empty;
+
+    [DataType(DataType.PhoneNumber)]
+    [Required(ErrorMessage = "Phone number is required")]
+    [CustomPhoneNumber(ErrorMessage = Constants.PhoneErrorMessage)]
+    [DisplayFormat(DataFormatString = Constants.PhoneNumberFormat)]
+    public string Phone { get; set; } = string.Empty;
+
+    [DataType(DataType.EmailAddress)]
+    [Required(ErrorMessage = "Email is required")]
+    [MaxLength(256)]
+    public string Email { get; set; } = string.Empty;
+
+    [Required(ErrorMessage = "Children's min age is required")]
+    [Range(0, 120, ErrorMessage = "Min age should be a number from 0 to 120")]
+    public int MinAge { get; set; }
+
+    [Required(ErrorMessage = "Children's max age is required")]
+    [Range(0, 120, ErrorMessage = "Max age should be a number from 0 to 120")]
+    public int MaxAge { get; set; }
+
+    [ModelBinder(BinderType = typeof(JsonModelBinder))]
+    [CollectionNotEmpty(ErrorMessage = "At least one DateTime range is required")]
+    public List<DateTimeRangeDto> DateTimeRanges { get; set; }
+
+    [Required(ErrorMessage = "Form of learning is required")]
+    [EnumDataType(typeof(FormOfLearning), ErrorMessage = Constants.EnumErrorMessage)]
+    public FormOfLearning FormOfLearning { get; set; } = FormOfLearning.Offline;
+
+    [Required(ErrorMessage = "Property IsPaid is required")]
+    public bool IsPaid { get; set; } = false;
+
+    [Column(TypeName = "decimal(18,2)")]
+    [Range(0, 100000, ErrorMessage = "Field value should be in a range from 1 to 100 000")]
+    public decimal? Price { get; set; } = default;
+
+    [EnumDataType(typeof(PayRateType), ErrorMessage = Constants.EnumErrorMessage)]
+    public PayRateType? PayRate { get; set; } = PayRateType.Classes;
+
+    [Required(ErrorMessage = "Available seats are required")]
+    public uint? AvailableSeats { get; set; } = uint.MaxValue;
+
+    [Required(ErrorMessage = "Property CompetitiveSelection is required")]
+    public bool CompetitiveSelection { get; set; }
+
+    [MaxLength(500)]
+    public string CompetitiveSelectionDescription { get; set; }
+
+    [Required]
+    public Guid ProviderId { get; set; }
+
+    public IEnumerable<ValidationResult> Validate(ValidationContext validationContext)
+    {
+        // TODO: Validate DateTimeRanges are not empty when frontend is ready
+        foreach (var dateTimeRange in DateTimeRanges)
+        {
+            if (dateTimeRange.StartTime > dateTimeRange.EndTime)
+            {
+                yield return new ValidationResult(
+                    "End date can't be earlier that start date");
+            }
+
+            if (dateTimeRange.Workdays.IsNullOrEmpty() || dateTimeRange.Workdays.Any(workday => workday == DaysBitMask.None))
+            {
+                yield return new ValidationResult(
+                    "Workdays are required");
+            }
+
+            var daysHs = new HashSet<DaysBitMask>();
+            if (!dateTimeRange.Workdays.All(daysHs.Add))
+            {
+                yield return new ValidationResult(
+                    "Workdays contain duplications");
+            }
+        }
+    }
+}
