+++ OutOfSchool/OutOfSchool.BusinessLogic/Models/Workshops/Drafts/WorkshopStaffDto.cs
@@ -1,0 +1,15 @@
+﻿using Microsoft.AspNetCore.Mvc;
+using OutOfSchool.BusinessLogic.Util.JsonTools;
+
+namespace OutOfSchool.BusinessLogic.Models.Workshops.Drafts;
+
+public class WorkshopStaffDto : WorkshopContactsDto
+{
+    [ModelBinder(BinderType = typeof(JsonModelBinder))]
+    public TeacherDTO DefaultTeacher { get; set; }
+
+    [ModelBinder(BinderType = typeof(JsonModelBinder))]
+    public List<TeacherDTO> Teachers { get; set; }
+
+    public Guid? DefaultTeacherId { get; set; }
+}
