+++ OutOfSchool/OutOfSchool.BusinessLogic/Models/Workshops/WorkshopCreateRequestDto.cs
@@ -1,190 +1,7 @@
-﻿using Microsoft.AspNetCore.Mvc;
-using OutOfSchool.BusinessLogic.Util.CustomValidation;
-using OutOfSchool.BusinessLogic.Util.JsonTools;
-using OutOfSchool.Common.Enums;
-using OutOfSchool.Common.Validators;
-using OutOfSchool.Services.Enums;
-using System.ComponentModel.DataAnnotations.Schema;
-using System.ComponentModel.DataAnnotations;
+﻿using OutOfSchool.BusinessLogic.Models.Workshops.Drafts;
 
 namespace OutOfSchool.BusinessLogic.Models.Workshops;
 
-public class WorkshopCreateRequestDto : IValidatableObject
+public class WorkshopCreateRequestDto : WorkshopStaffDto
 {
-    public Guid Id { get; set; }
-
-    [Required(ErrorMessage = "Workshop title is required")]
-    [MinLength(Constants.MinWorkshopTitleLength)]
-    [MaxLength(Constants.MaxWorkshopTitleLength)]
-    public string Title { get; set; } = string.Empty;
-
-    [Required(ErrorMessage = "Workshop short title is required")]
-    [MinLength(Constants.MinWorkshopShortTitleLength)]
-    [MaxLength(Constants.MaxWorkshopShortTitleLength)]
-    public string ShortTitle { get; set; } = string.Empty;
-
-    [DataType(DataType.PhoneNumber)]
-    [Required(ErrorMessage = "Phone number is required")]
-    [CustomPhoneNumber(ErrorMessage = Constants.PhoneErrorMessage)]
-    [DisplayFormat(DataFormatString = Constants.PhoneNumberFormat)]
-    public string Phone { get; set; } = string.Empty;
-
-    [DataType(DataType.EmailAddress)]
-    [Required(ErrorMessage = "Email is required")]
-    [MaxLength(256)]
-    public string Email { get; set; } = string.Empty;
-
-    [DataType(DataType.Url)]
-    [MaxLength(Constants.MaxUnifiedUrlLength)]
-    public string Website { get; set; } = string.Empty;
-
-    [DataType(DataType.Url)]
-    [MaxLength(Constants.MaxUnifiedUrlLength)]
-    public string Facebook { get; set; } = string.Empty;
-
-    [DataType(DataType.Url)]
-    [MaxLength(Constants.MaxUnifiedUrlLength)]
-    public string Instagram { get; set; } = string.Empty;
-
-    [Required(ErrorMessage = "Children's min age is required")]
-    [Range(0, 120, ErrorMessage = "Min age should be a number from 0 to 120")]
-    public int MinAge { get; set; }
-
-    [Required(ErrorMessage = "Children's max age is required")]
-    [Range(0, 120, ErrorMessage = "Max age should be a number from 0 to 120")]
-    public int MaxAge { get; set; }
-
-    [ModelBinder(BinderType = typeof(JsonModelBinder))]
-    [CollectionNotEmpty(ErrorMessage = "At least one DateTime range is required")]
-    public List<DateTimeRangeDto> DateTimeRanges { get; set; }
-
-    public bool IsPaid { get; set; } = false;
-
-    [Column(TypeName = "decimal(18,2)")]
-    [Range(0, 100000, ErrorMessage = "Field value should be in a range from 1 to 100 000")]
-    public decimal? Price { get; set; } = default;
-
-    [EnumDataType(typeof(PayRateType), ErrorMessage = Constants.EnumErrorMessage)]
-    public PayRateType? PayRate { get; set; } = PayRateType.Classes;
-
-    [Required(ErrorMessage = "Form of learning is required")]
-    [EnumDataType(typeof(FormOfLearning), ErrorMessage = Constants.EnumErrorMessage)]
-    public FormOfLearning FormOfLearning { get; set; } = FormOfLearning.Offline;
-
-    [Required(ErrorMessage = "Available seats are required")]
-    public uint? AvailableSeats { get; set; } = uint.MaxValue;
-
-    public bool CompetitiveSelection { get; set; }
-
-    [MaxLength(500)]
-    public string CompetitiveSelectionDescription { get; set; }
-
-    [ModelBinder(BinderType = typeof(JsonModelBinder))]
-    [CollectionNotEmpty(ErrorMessage = "At least one description item is required")]
-    public IEnumerable<WorkshopDescriptionItemDto> WorkshopDescriptionItems { get; set; }
-
-    public bool WithDisabilityOptions { get; set; } = default;
-
-    [MaxLength(200)]
-    public string DisabilityOptionsDesc { get; set; } = string.Empty;
-
-    public Guid? InstitutionId { get; set; }
-
-    public Guid? InstitutionHierarchyId { get; set; }
-
-    [ModelBinder(BinderType = typeof(JsonModelBinder))]
-    public TeacherDTO DefaultTeacher { get; set; }
-
-    public List<long> DirectionIds { get; set; }
-
-    [ModelBinder(BinderType = typeof(JsonModelBinder))]
-    public IEnumerable<string> Keywords { get; set; } = default;
-
-    [ModelBinder(BinderType = typeof(JsonModelBinder))]
-    public List<TeacherDTO> Teachers { get; set; }
-
-    [Required]
-    public Guid ProviderId { get; set; }
-
-    public DateOnly ActiveFrom { get; set; }
-
-    public DateOnly ActiveTo { get; set; }
-
-    public bool ShortStay { get; set; } = false;
-
-    public bool IsSelfFinanced { get; set; } = false;
-
-    public bool IsSpecial { get; set; } = false;
-
-    public uint SpecialNeedsId { get; set; } = 0;
-
-    public bool IsInclusive { get; set; } = false;
-
-    [MaxLength(500)]
-    public string AdditionalDescription { get; set; }
-
-    public bool AreThereBenefits { get; set; } = default;
-
-    [MaxLength(500)]
-    public string PreferentialTermsOfParticipation { get; set; }
-
-    [Required]
-    public uint EducationalShiftId { get; set; } = 0;
-
-    [Required(ErrorMessage = "Language of education is required")]
-    public uint LanguageOfEducationId { get; set; } = 0;
-
-    [Required(ErrorMessage = "Type of age composition is required")]
-    public uint TypeOfAgeCompositionId { get; set; } = 0;
-
-    [Required(ErrorMessage = "Educational disciplines is required")]
-    public Guid EducationalDisciplines { get; set; } = Guid.Empty;
-
-    [Required(ErrorMessage = "Category is required")]
-    public uint CategoryId { get; set; } = 0;
-
-    [Required(ErrorMessage = "GropeType is required")]
-    public uint GropeTypeId { get; set; } = 0;
-
-    public uint CoverageId { get; set; } = 0;
-
-    public Guid? DefaultTeacherId { get; set; }
-
-    public Guid? MemberOfWorkshopId { get; set; }
-
-    [Required]
-    public long AddressId { get; set; }
-
-    [Required]
-    [ModelBinder(BinderType = typeof(JsonModelBinder))]
-    public AddressDto Address { get; set; }
-
-    [ModelBinder(BinderType = typeof(JsonModelBinder))]
-    public List<long> TagIds { get; set; } = [];
-
-    public IEnumerable<ValidationResult> Validate(ValidationContext validationContext)
-    {
-        // TODO: Validate DateTimeRanges are not empty when frontend is ready
-        foreach (var dateTimeRange in DateTimeRanges)
-        {
-            if (dateTimeRange.StartTime > dateTimeRange.EndTime)
-            {
-                yield return new ValidationResult(
-                    "End date can't be earlier that start date");
-            }
-
-            if (dateTimeRange.Workdays.IsNullOrEmpty() || dateTimeRange.Workdays.Any(workday => workday == DaysBitMask.None))
-            {
-                yield return new ValidationResult(
-                    "Workdays are required");
-            }
-
-            var daysHs = new HashSet<DaysBitMask>();
-            if (!dateTimeRange.Workdays.All(daysHs.Add))
-            {
-                yield return new ValidationResult(
-                    "Workdays contain duplications");
-            }
-        }
-    }
 }
